feat: getAllCategories method added fix: the removeCategory also catches if a category is used by product when you want to remove

+++ src/main/java/gr/aueb/cf/e_commerce_app/service/CategoryService.java
@@ -1,6 +1,7 @@
 package gr.aueb.cf.e_commerce_app.service;
 
 import gr.aueb.cf.e_commerce_app.core.exceptions.AppObjectAlreadyExistsException;
+import gr.aueb.cf.e_commerce_app.core.exceptions.AppObjectIllegalStateException;
 import gr.aueb.cf.e_commerce_app.core.exceptions.AppObjectNotFoundException;
 import gr.aueb.cf.e_commerce_app.dto.CategoryInsertDto;
 import gr.aueb.cf.e_commerce_app.dto.CategoryReadOnlyDto;
@@ -9,7 +10,10 @@ import gr.aueb.cf.e_commerce_app.model.static_data.Category;
 import gr.aueb.cf.e_commerce_app.repository.CategoryRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -31,10 +35,19 @@ public class CategoryService {
     }
 
     @Transactional(rollbackOn = Exception.class)
-    public void removeCategory(Long id) throws AppObjectNotFoundException {
+    public void removeCategory(Long id) throws AppObjectNotFoundException, AppObjectIllegalStateException {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new AppObjectNotFoundException("Category", "The category you want to remove is not found"));
 
+        if (!category.getAllProducts().isEmpty()) throw new AppObjectIllegalStateException("Category", "You cannot remove a category that is used by products");
+
         categoryRepository.delete(category);
+    }
+
+    public List<CategoryReadOnlyDto> getAllCategories() {
+
+        return categoryRepository.findAll(Sort.by("id")).stream()
+                .map(mapper::mapToCategoryReadOnlyDto)
+                .toList();
     }
 }
